feat: signup page refactor
