+++ packages/frontend/app/auth/signup/page.tsx
@@ -1,1131 +1,103 @@
 'use client';
 
-import { AuthService } from '@/api/services/auth.service';
-import CommonInput from '@/components/atoms/CommonInput';
-import EmailDisplayBox from '@/components/atoms/EmailDisplayBox';
-import PasswordInputGroup from '@/components/atoms/PasswordInputGroup';
-import InterestRegionSelect from '@/components/atoms/select/InterestRegionSelect';
-import SignupCompleteModal from '@/components/ui/SignupCompleteModal';
-import {
-    validateConfirmPassword,
-    validateName,
-    validateOrganization,
-    validatePassword,
-    validatePhoneNumber,
-    validateSignupForm,
-} from '@/utils/validation';
-import { useRouter } from 'next/navigation';
-import { useState } from 'react';
-
-// UserType 타입 정의
-type UserType = 'GOV' | 'EDU' | 'GENERAL';
+import { useSignupFlow } from '@/features/signup/hooks/useSignupFlow';
+import CodeVerificationSection from '@/features/signup/sections/CodeVerificationSection';
+import EmailVerificationSection from '@/features/signup/sections/EmailVerificationSection';
+import ProfileSection from '@/features/signup/sections/ProfileSection';
 
 export default function SignUpPage() {
-  const router = useRouter();
-  
-  // 1/2 단계 상태
-  const [email, setEmail] = useState('');
-  const [verificationCode, setVerificationCode] = useState('');
-  const [isLoading, setIsLoading] = useState(false);
-  const [isVerifying, setIsVerifying] = useState(false);
-  const [isCodeSent, setIsCodeSent] = useState(false);
-  const [error, setError] = useState('');
-  const [verificationError, setVerificationError] = useState('');
-  const [countdown, setCountdown] = useState(0);
-  
-  // 2/2 단계 상태
-  const [isVerified, setIsVerified] = useState(false);
-  const [isProfileLoading, setIsProfileLoading] = useState(false);
-  const [showCompleteModal, setShowCompleteModal] = useState(false);
-  const [userType, setUserType] = useState<UserType>('GENERAL');
-  const [formData, setFormData] = useState({
-    name: '',
-    password: '',
-    confirmPassword: '',
-    organization: '',
-    phoneNumber: '',
-    regionId: '',
-    agreeToAge: false,
-    agreeToTerms: false,
-    agreeToPrivacy: false,
-    agreeToMarketing: false,
-    agreeToReportReservation: false,
-  });
-  const [errors, setErrors] = useState({
-    name: '',
-    password: '',
-    confirmPassword: '',
-    organization: '',
-    phoneNumber: '',
-  });
-  const [touched, setTouched] = useState({
-    name: false,
-    password: false,
-    confirmPassword: false,
-    organization: false,
-    phoneNumber: false,
-  });
+  const {
+    // State
+    email,
+    setEmail,
+    verificationCode,
+    setVerificationCode,
+    isLoading,
+    isVerifying,
+    isCodeSent,
+    error,
+    verificationError,
+    countdown,
+    isVerified,
+    isProfileLoading,
+    showCompleteModal,
+    userType,
+    formData,
+    setFormData,
+    
+    // Handlers
+    handleEmailSubmit,
+    handleCodeSubmit,
+    handleResendCode,
+    handleBackToStep1,
+    handleSubmit,
+    handleStart,
+    formatTime,
+  } = useSignupFlow();
 
-  const reportReservationLink = 'https://dev.klaci.kr';
-  const termsLink = 'https://dev.klaci.kr';
-  const privacyLink = 'https://dev.klaci.kr';
-  const marketingLink = 'https://dev.klaci.kr';
-
-  // 카운트다운 타이머
-  const startCountdown = () => {
-    setCountdown(180); // 3분
-    const timer = setInterval(() => {
-      setCountdown((prev) => {
-        if (prev <= 1) {
-          clearInterval(timer);
-          return 0;
-        }
-        return prev - 1;
-      });
-    }, 1000);
-  };
-
-  // 인증번호 발송
-  const handleEmailSubmit = async (e: React.FormEvent) => {
-    e.preventDefault();
-    setIsLoading(true);
-    setError('');
-    setVerificationError('');
-
-    try {
-      await AuthService.sendVerificationCode({ email });
-      setIsCodeSent(true);
-      startCountdown();
-    } catch (err: any) {
-      setError(
-        err.response?.data?.message || '인증번호 발송 중 오류가 발생했습니다.',
-      );
-    } finally {
-      setIsLoading(false);
-    }
-  };
-
-  // 인증번호 검증
-  const handleCodeSubmit = async (e: React.FormEvent) => {
-    e.preventDefault();
-    setIsVerifying(true);
-    setVerificationError('');
-
-    try {
-      const response = await AuthService.verifyCode({ email, code: verificationCode });
-      if (response.data.verified) {
-        // 인증 성공 - 2/2 단계로 전환
-        setIsVerified(true);
-        // 이메일 도메인에 따른 사용자 타입 설정
-        const domain = email.split('@')[1];
-        if (domain?.includes('korea.kr') || domain?.includes('go.kr')) {
-          setUserType('GOV');
-        } else if (domain?.includes('ac.kr')) {
-          setUserType('EDU');
-        } else {
-          setUserType('GENERAL');
-        }
-      } else {
-        setVerificationError('인증번호가 일치하지 않습니다.');
-      }
-    } catch (err: any) {
-      setVerificationError(
-        err.response?.data?.message || '인증번호 검증 중 오류가 발생했습니다.',
-      );
-    } finally {
-      setIsVerifying(false);
-    }
-  };
-
-  // 인증번호 재발송
-  const handleResendCode = async () => {
-    setIsLoading(true);
-    setError('');
-    setVerificationError('');
-
-    try {
-      await AuthService.sendVerificationCode({ email });
-      startCountdown();
-    } catch (err: any) {
-      setError(
-        err.response?.data?.message || '인증번호 재발송 중 오류가 발생했습니다.',
-      );
-    } finally {
-      setIsLoading(false);
-    }
-  };
-
-  // 1/2 단계로 돌아가기
-  const handleBackToStep1 = () => {
-    setIsCodeSent(false);
-    setIsVerified(false);
-    setVerificationCode('');
-    setCountdown(0);
-    setError('');
-    setVerificationError('');
-  };
-
-  // 2/2 단계 블러 이벤트 핸들러들
-  const handleNameBlur = () => {
-    setTouched((prev) => ({ ...prev, name: true }));
-    const validation = validateName(formData.name);
-    setErrors((prev) => ({ ...prev, name: validation.message }));
-  };
-
-  const handlePasswordBlur = () => {
-    setTouched((prev) => ({ ...prev, password: true }));
-    const validation = validatePassword(formData.password);
-    setErrors((prev) => ({ ...prev, password: validation.message }));
-  };
-
-  const handleConfirmPasswordBlur = () => {
-    setTouched((prev) => ({ ...prev, confirmPassword: true }));
-    const validation = validateConfirmPassword(
-      formData.confirmPassword,
-      formData.password,
-    );
-    setErrors((prev) => ({ ...prev, confirmPassword: validation.message }));
-  };
-
-  const handleOrganizationBlur = () => {
-    setTouched((prev) => ({ ...prev, organization: true }));
-    const validation = validateOrganization(formData.organization);
-    setErrors((prev) => ({ ...prev, organization: validation.message }));
-  };
-
-  const handlePhoneNumberBlur = () => {
-    setTouched((prev) => ({ ...prev, phoneNumber: true }));
-    const validation = validatePhoneNumber(formData.phoneNumber);
-    setErrors((prev) => ({ ...prev, phoneNumber: validation.message }));
-  };
-
-  // 회원가입 완료
-  const handleSubmit = async (e: React.FormEvent) => {
-    e.preventDefault();
-    setIsProfileLoading(true);
-    setError('');
-
-    // 필수 항목 검증
-    const validationResult = validateSignupForm(formData);
-    if (!validationResult.isValid) {
-      setError(validationResult.message);
-      setIsProfileLoading(false);
-      return;
-    }
-
-    try {
-      const signupData = {
-        email: email,
-        password: formData.password,
-        name: formData.name,
-        phone_number: formData.phoneNumber,
-        interest_region_id: formData.regionId || null,
-        organization: formData.organization,
-        user_type: userType,
-        agree_to_age: formData.agreeToAge,
-        agree_to_terms: formData.agreeToTerms,
-        agree_to_privacy: formData.agreeToPrivacy,
-        agree_to_marketing: formData.agreeToMarketing,
-        agree_to_report_reservation: formData.agreeToReportReservation,
-      };
-
-      console.log('Submitting signup data:', signupData);
-
-      const response = await AuthService.signUp(signupData);
-
-      console.log('Signup response:', response);
-      setShowCompleteModal(true);
-    } catch (err: any) {
-      console.error('Signup error:', err);
-      setError(
-        err.response?.data?.message || '회원가입 중 오류가 발생했습니다.',
-      );
-    } finally {
-      setIsProfileLoading(false);
-    }
-  };
-
-  const handleStart = () => {
-    setShowCompleteModal(false);
-    // 메인 페이지로 이동
-    router.push('/');
-  };
-
+  // 링크 핸들러들
   const handleReportReservationLink = () => {
-    window.open(reportReservationLink, '_blank');
+    window.open('https://dev.klaci.kr', '_blank');
   };
 
   const handleTermsLink = () => {
-    window.open(termsLink, '_blank');
+    window.open('https://dev.klaci.kr', '_blank');
   };
 
   const handlePrivacyLink = () => {
-    window.open(privacyLink, '_blank');
+    window.open('https://dev.klaci.kr', '_blank');
   };
 
   const handleMarketingLink = () => {
-    window.open(marketingLink, '_blank');
-  };
-
-  const descriptionTexts = [
-    'KLACI 계정으로 사용할 이메일 인증이 필요합니다.',
-    '이메일 도메인에 따라 회원 유형이 자동 분류되며 맞춤 정보가 제공됩니다.',
-  ];
-
-  const formatTime = (seconds: number) => {
-    const mins = Math.floor(seconds / 60);
-    const secs = seconds % 60;
-    return `${mins}:${secs.toString().padStart(2, '0')}`;
+    window.open('https://dev.klaci.kr', '_blank');
   };
 
   // 2/2 단계 렌더링
   if (isVerified) {
     return (
-      <div
-        style={{
-          width: '100%',
-          display: 'flex',
-          alignItems: 'center',
-          justifyContent: 'center',
-          backgroundColor: '#F4F4F4',
-          paddingTop: '80px',
-        }}
-      >
-        <div style={{ width: '100%', maxWidth: '800px', margin: '0 1rem' }}>
-          {/* 메인 컨테이너 */}
-          <div
-            style={{
-              display: 'flex',
-              flexDirection: 'column',
-              alignItems: 'center',
-              justifyContent: 'center',
-              backgroundColor: 'white',
-              borderRadius: '50px',
-              padding: '80px 100px',
-            }}
-          >
-            {/* 2/2 단계 뱃지 */}
-            <div
-              style={{
-                display: 'flex',
-                justifyContent: 'center',
-                marginBottom: '2rem',
-              }}
-            >
-              <div
-                style={{
-                  border: '1px solid black',
-                  borderRadius: '12px',
-                  padding: '6px 12px',
-                  fontSize: '16px',
-                  fontWeight: '500',
-                  color: 'black',
-                }}
-              >
-                2/2 단계
-              </div>
-            </div>
-
-            {/* 타이틀 */}
-            <div style={{ textAlign: 'center', marginBottom: '2rem' }}>
-              <h1
-                style={{
-                  fontSize: '1.875rem',
-                  fontWeight: '700',
-                  color: '#111827',
-                }}
-              >
-                회원정보 입력
-              </h1>
-            </div>
-
-            {/* 폼 */}
-            <form
-              onSubmit={handleSubmit}
-              style={{
-                width: '70%',
-                display: 'flex',
-                flexDirection: 'column',
-                alignItems: 'center',
-                justifyContent: 'center',
-                gap: '20px',
-              }}
-            >
-              {/* 이메일 표시 박스 */}
-              <EmailDisplayBox email={email} userType={userType} />
-
-              {/* 이름 입력 */}
-              <CommonInput
-                id="name"
-                label="이름"
-                value={formData.name}
-                onChange={(e) =>
-                  setFormData({ ...formData, name: e.target.value })
-                }
-                placeholder="한글 또는 영어 10자 이하"
-                required={true}
-                maxLength={10}
-                isRequired={true}
-                error={touched.name ? errors.name : ''}
-                onBlur={handleNameBlur}
-              />
-
-              {/* 비밀번호 입력 그룹 */}
-              <PasswordInputGroup
-                password={formData.password}
-                confirmPassword={formData.confirmPassword}
-                onPasswordChange={(e) =>
-                  setFormData({ ...formData, password: e.target.value })
-                }
-                onConfirmPasswordChange={(e) =>
-                  setFormData({ ...formData, confirmPassword: e.target.value })
-                }
-                passwordError={touched.password ? errors.password : ''}
-                confirmPasswordError={
-                  touched.confirmPassword ? errors.confirmPassword : ''
-                }
-                onPasswordBlur={handlePasswordBlur}
-                onConfirmPasswordBlur={handleConfirmPasswordBlur}
-              />
-
-              {/* 소속 입력 */}
-              <CommonInput
-                id="organization"
-                label="소속"
-                value={formData.organization}
-                onChange={(e) =>
-                  setFormData({ ...formData, organization: e.target.value })
-                }
-                placeholder="한글 또는 영어 10자 이하"
-                required={true}
-                maxLength={10}
-                isRequired={true}
-                error={touched.organization ? errors.organization : ''}
-                onBlur={handleOrganizationBlur}
-              />
-
-              {/* 휴대폰 번호 입력 */}
-              <CommonInput
-                id="phoneNumber"
-                label="휴대폰 번호"
-                value={formData.phoneNumber}
-                onChange={(e) =>
-                  setFormData({ ...formData, phoneNumber: e.target.value })
-                }
-                placeholder="숫자만 입력"
-                type="tel"
-                required={true}
-                isRequired={true}
-                error={touched.phoneNumber ? errors.phoneNumber : ''}
-                onBlur={handlePhoneNumberBlur}
-              />
-
-              {/* 관심 지역 설정 */}
-              <div style={{ width: '100%' }}>
-                <div
-                  style={{
-                    display: 'flex',
-                    justifyContent: 'space-between',
-                    alignItems: 'center',
-                    marginBottom: '8px',
-                  }}
-                >
-                  <label
-                    style={{
-                      fontSize: '14px',
-                      fontWeight: '500',
-                      color: '#374151',
-                    }}
-                  >
-                    관심 지역 설정
-                  </label>
-                  <span
-                    style={{
-                      fontSize: '12px',
-                      color: '#9A9EA3',
-                      fontWeight: '500',
-                    }}
-                  >
-                    선택 사항
-                  </span>
-                </div>
-                <InterestRegionSelect
-                  selectedRegionId={formData.regionId}
-                  onRegionChange={(regionId) => {
-                    setFormData((prev) => ({ ...prev, regionId }));
-                  }}
-                />
-              </div>
-
-              {/* KLACI 인사이트 리포트 사전예약 신청 */}
-              <div style={{ width: '100%' }}>
-                <div
-                  style={{
-                    display: 'flex',
-                    justifyContent: 'space-between',
-                    alignItems: 'center',
-                    marginBottom: '8px',
-                  }}
-                >
-                  <div style={{ display: 'flex', alignItems: 'center' }}>
-                    <input
-                      type="checkbox"
-                      id="agreeToReportReservation"
-                      checked={formData.agreeToReportReservation}
-                      onChange={(e) =>
-                        setFormData({
-                          ...formData,
-                          agreeToReportReservation: e.target.checked,
-                        })
-                      }
-                      style={{
-                        marginRight: '8px',
-                      }}
-                    />
-                    <label
-                      htmlFor="agreeToReportReservation"
-                      style={{
-                        fontSize: '14px',
-                        fontWeight: '700',
-                        color: '#374151',
-                      }}
-                    >
-                      KLACI 인사이트 리포트 사전예약 신청
-                    </label>
-                  </div>
-                  <span
-                    style={{
-                      fontSize: '12px',
-                      color: '#9A9EA3',
-                      fontWeight: '500',
-                    }}
-                  >
-                    선택 사항
-                  </span>
-                </div>
-
-                {/* 설명 텍스트 */}
-                <div
-                  style={{
-                    fontSize: '12px',
-                    color: '#6B7280',
-                    marginBottom: '4px',
-                    marginLeft: '24px',
-                  }}
-                >
-                  사전예약자에 한해 특별 할인코드를 메일로 보내드립니다
-                </div>
-
-                {/* 더 알아보기 링크 */}
-                <div style={{ marginLeft: '24px' }}>
-                  <button
-                    type="button"
-                    onClick={handleReportReservationLink}
-                    style={{
-                      background: 'none',
-                      border: 'none',
-                      padding: 0,
-                      fontSize: '12px',
-                      color: '#000000',
-                      textDecoration: 'underline',
-                      cursor: 'pointer',
-                      fontFamily: 'inherit',
-                    }}
-                  >
-                    더 알아보기 (새창)
-                  </button>
-                </div>
-              </div>
-
-              {/* 동의 항목 박스 */}
-              <div
-                style={{
-                  width: '100%',
-                  backgroundColor: '#F1F1F1',
-                  borderRadius: '0.5rem',
-                  padding: '20px',
-                  marginTop: '20px',
-                  marginBottom: '20px',
-                }}
-              >
-                {/* 만 14세 이상 */}
-                <div
-                  style={{
-                    display: 'flex',
-                    alignItems: 'center',
-                    marginBottom: '12px',
-                  }}
-                >
-                  <input
-                    type="checkbox"
-                    id="agreeToAge"
-                    checked={formData.agreeToAge}
-                    onChange={(e) =>
-                      setFormData({ ...formData, agreeToAge: e.target.checked })
-                    }
-                    style={{
-                      marginRight: '8px',
-                    }}
-                  />
-                  <label
-                    htmlFor="agreeToAge"
-                    style={{
-                      fontSize: '14px',
-                      color: '#374151',
-                    }}
-                  >
-                    만 14세 이상입니다
-                  </label>
-                </div>
-
-                {/* 서비스 이용약관 */}
-                <div
-                  style={{
-                    display: 'flex',
-                    alignItems: 'center',
-                    marginBottom: '12px',
-                  }}
-                >
-                  <input
-                    type="checkbox"
-                    id="agreeToServiceTerms"
-                    checked={formData.agreeToTerms}
-                    onChange={(e) =>
-                      setFormData({
-                        ...formData,
-                        agreeToTerms: e.target.checked,
-                      })
-                    }
-                    style={{
-                      marginRight: '8px',
-                    }}
-                  />
-                  <label
-                    htmlFor="agreeToServiceTerms"
-                    style={{
-                      fontSize: '14px',
-                      color: '#374151',
-                    }}
-                  >
-                    <button
-                      type="button"
-                      onClick={handleTermsLink}
-                      style={{
-                        background: 'none',
-                        border: 'none',
-                        padding: 0,
-                        fontSize: '14px',
-                        color: '#374151',
-                        textDecoration: 'underline',
-                        cursor: 'pointer',
-                        fontFamily: 'inherit',
-                      }}
-                    >
-                      서비스 이용약관
-                    </button>
-                  </label>
-                </div>
-
-                {/* 개인정보 수집 및 이용 */}
-                <div
-                  style={{
-                    display: 'flex',
-                    alignItems: 'center',
-                    marginBottom: '12px',
-                  }}
-                >
-                  <input
-                    type="checkbox"
-                    id="agreeToPrivacy"
-                    checked={formData.agreeToPrivacy}
-                    onChange={(e) =>
-                      setFormData({
-                        ...formData,
-                        agreeToPrivacy: e.target.checked,
-                      })
-                    }
-                    style={{
-                      marginRight: '8px',
-                    }}
-                  />
-                  <label
-                    htmlFor="agreeToPrivacy"
-                    style={{
-                      fontSize: '14px',
-                      color: '#374151',
-                    }}
-                  >
-                    <button
-                      type="button"
-                      onClick={handlePrivacyLink}
-                      style={{
-                        background: 'none',
-                        border: 'none',
-                        padding: 0,
-                        fontSize: '14px',
-                        color: '#374151',
-                        textDecoration: 'underline',
-                        cursor: 'pointer',
-                        fontFamily: 'inherit',
-                      }}
-                    >
-                      개인정보 수집 및 이용
-                    </button>
-                  </label>
-                </div>
-
-                {/* 세미나 및 이벤트 등 마케팅 정보 수신 */}
-                <div
-                  style={{
-                    display: 'flex',
-                    alignItems: 'center',
-                  }}
-                >
-                  <input
-                    type="checkbox"
-                    id="agreeToMarketing"
-                    checked={formData.agreeToMarketing}
-                    onChange={(e) =>
-                      setFormData({
-                        ...formData,
-                        agreeToMarketing: e.target.checked,
-                      })
-                    }
-                    style={{
-                      marginRight: '8px',
-                    }}
-                  />
-                  <label
-                    htmlFor="agreeToMarketing"
-                    style={{
-                      fontSize: '14px',
-                      color: '#374151',
-                    }}
-                  >
-                    <button
-                      type="button"
-                      onClick={handleMarketingLink}
-                      style={{
-                        background: 'none',
-                        border: 'none',
-                        padding: 0,
-                        fontSize: '14px',
-                        color: '#374151',
-                        textDecoration: 'underline',
-                        cursor: 'pointer',
-                        fontFamily: 'inherit',
-                      }}
-                    >
-                      세미나 및 이벤트 등 마케팅 정보 수신
-                    </button>
-                    <span
-                      style={{
-                        fontSize: '14px',
-                        color: '#6B7280',
-                        marginLeft: '4px',
-                      }}
-                    >
-                      (선택)
-                    </span>
-                  </label>
-                </div>
-              </div>
-
-              {/* 회원가입 완료 버튼 */}
-              <button
-                type="submit"
-                disabled={isProfileLoading}
-                style={{
-                  width: '100%',
-                  height: '50px',
-                  backgroundColor: isProfileLoading ? '#9CA3AF' : '#000000',
-                  color: 'white',
-                  borderRadius: '0.5rem',
-                  fontWeight: '500',
-                  cursor: isProfileLoading ? 'not-allowed' : 'pointer',
-                  border: 'none',
-                  transition: 'background-color 0.2s',
-                  fontSize: '16px',
-                }}
-                onMouseEnter={(e) => {
-                  if (!isProfileLoading) {
-                    e.currentTarget.style.backgroundColor = '#1F2937';
-                  }
-                }}
-                onMouseLeave={(e) => {
-                  if (!isProfileLoading) {
-                    e.currentTarget.style.backgroundColor = '#000000';
-                  }
-                }}
-              >
-                {isProfileLoading ? (
-                  <div
-                    style={{
-                      display: 'flex',
-                      alignItems: 'center',
-                      justifyContent: 'center',
-                    }}
-                  >
-                    <div
-                      style={{
-                        animation: 'spin 1s linear infinite',
-                        borderRadius: '50%',
-                        height: '1rem',
-                        width: '1rem',
-                        border: '2px solid transparent',
-                        borderBottomColor: 'white',
-                        marginRight: '0.5rem',
-                      }}
-                    ></div>
-                    처리 중...
-                  </div>
-                ) : (
-                  '회원가입 완료'
-                )}
-              </button>
-            </form>
-          </div>
-        </div>
-
-        {/* 회원가입 완료 모달 */}
-        <SignupCompleteModal
-          isOpen={showCompleteModal}
-          onClose={() => setShowCompleteModal(false)}
-          onStart={handleStart}
-        />
-      </div>
+      <ProfileSection
+        email={email}
+        userType={userType}
+        formData={formData}
+        setFormData={setFormData}
+        isProfileLoading={isProfileLoading}
+        showCompleteModal={showCompleteModal}
+        onSubmit={handleSubmit}
+        onStart={handleStart}
+        onReportReservationLink={handleReportReservationLink}
+        onTermsLink={handleTermsLink}
+        onPrivacyLink={handlePrivacyLink}
+        onMarketingLink={handleMarketingLink}
+        error={error}
+      />
     );
   }
 
   // 1/2 단계 렌더링
+  if (isCodeSent) {
+    return (
+      <CodeVerificationSection
+        verificationCode={verificationCode}
+        setVerificationCode={setVerificationCode}
+        isVerifying={isVerifying}
+        isLoading={isLoading}
+        countdown={countdown}
+        onSubmit={handleCodeSubmit}
+        onResend={handleResendCode}
+        onBackToStep1={handleBackToStep1}
+        formatTime={formatTime}
+        verificationError={verificationError}
+      />
+    );
+  }
+
   return (
-    <div
-      style={{
-        width: '100%',
-        display: 'flex',
-        alignItems: 'center',
-        justifyContent: 'center',
-        backgroundColor: '#F4F4F4',
-        paddingTop: '80px',
-      }}
-    >
-      <div style={{ width: '100%', maxWidth: '800px', margin: '0 1rem' }}>
-        {/* 메인 컨테이너 */}
-        <div
-          style={{
-            display: 'flex',
-            flexDirection: 'column',
-            alignItems: 'center',
-            justifyContent: 'center',
-            backgroundColor: 'white',
-            borderRadius: '50px',
-            padding: '80px 100px',
-          }}
-        >
-          {/* 1/2 단계 뱃지 */}
-          <div
-            style={{
-              display: 'flex',
-              justifyContent: 'center',
-              marginBottom: '2rem',
-            }}
-          >
-            <div
-              style={{
-                border: '1px solid black',
-                borderRadius: '12px',
-                padding: '6px 12px',
-                fontSize: '16px',
-                fontWeight: '500',
-                color: 'black',
-              }}
-            >
-              1/2 단계
-            </div>
-          </div>
-
-          {/* 타이틀 */}
-          <div style={{ textAlign: 'center', marginBottom: '2rem' }}>
-            <h1
-              style={{
-                fontSize: '1.875rem',
-                fontWeight: '700',
-                color: '#111827',
-              }}
-            >
-              이메일 인증
-            </h1>
-          </div>
-
-          {/* 설명 텍스트 */}
-          <div
-            style={{
-              textAlign: 'center',
-              marginBottom: '2rem',
-              maxWidth: '400px',
-            }}
-          >
-            {descriptionTexts.map((text, index) => (
-              <p
-                key={index}
-                style={{
-                  fontSize: '14px',
-                  color: '#6B7280',
-                  marginBottom: '0.5rem',
-                }}
-              >
-                {text}
-              </p>
-            ))}
-          </div>
-
-          {/* 폼 */}
-          {!isCodeSent ? (
-            <form
-              onSubmit={handleEmailSubmit}
-              style={{
-                width: '70%',
-                display: 'flex',
-                flexDirection: 'column',
-                alignItems: 'center',
-                justifyContent: 'center',
-                gap: '20px',
-              }}
-            >
-              {/* 이메일 입력 */}
-              <CommonInput
-                id="email"
-                label="이메일"
-                value={email}
-                onChange={(e) => setEmail(e.target.value)}
-                placeholder="이메일을 입력해주세요"
-                type="email"
-                required={true}
-                isRequired={true}
-              />
-
-              {/* 인증번호 발송 버튼 */}
-              <button
-                type="submit"
-                disabled={isLoading}
-                style={{
-                  width: '100%',
-                  height: '50px',
-                  backgroundColor: isLoading ? '#9CA3AF' : '#000000',
-                  color: 'white',
-                  borderRadius: '0.5rem',
-                  fontWeight: '500',
-                  cursor: isLoading ? 'not-allowed' : 'pointer',
-                  border: 'none',
-                  transition: 'background-color 0.2s',
-                  fontSize: '16px',
-                }}
-                onMouseEnter={(e) => {
-                  if (!isLoading) {
-                    e.currentTarget.style.backgroundColor = '#1F2937';
-                  }
-                }}
-                onMouseLeave={(e) => {
-                  if (!isLoading) {
-                    e.currentTarget.style.backgroundColor = '#000000';
-                  }
-                }}
-              >
-                {isLoading ? (
-                  <div
-                    style={{
-                      display: 'flex',
-                      alignItems: 'center',
-                      justifyContent: 'center',
-                    }}
-                  >
-                    <div
-                      style={{
-                        animation: 'spin 1s linear infinite',
-                        borderRadius: '50%',
-                        height: '1rem',
-                        width: '1rem',
-                        border: '2px solid transparent',
-                        borderBottomColor: 'white',
-                        marginRight: '0.5rem',
-                      }}
-                    ></div>
-                    처리 중...
-                  </div>
-                ) : (
-                  '인증번호 발송'
-                )}
-              </button>
-            </form>
-          ) : (
-            <form
-              onSubmit={handleCodeSubmit}
-              style={{
-                width: '70%',
-                display: 'flex',
-                flexDirection: 'column',
-                alignItems: 'center',
-                justifyContent: 'center',
-                gap: '20px',
-              }}
-            >
-              {/* 인증번호 입력 */}
-              <CommonInput
-                id="verificationCode"
-                label="인증번호"
-                value={verificationCode}
-                onChange={(e) => setVerificationCode(e.target.value)}
-                placeholder="인증번호를 입력해주세요"
-                required={true}
-                isRequired={true}
-              />
-
-              {/* 카운트다운 */}
-              {countdown > 0 && (
-                <div
-                  style={{
-                    fontSize: '14px',
-                    color: '#6B7280',
-                    textAlign: 'center',
-                  }}
-                >
-                  {formatTime(countdown)}
-                </div>
-              )}
-
-              {/* 인증번호 검증 버튼 */}
-              <button
-                type="submit"
-                disabled={isVerifying}
-                style={{
-                  width: '100%',
-                  height: '50px',
-                  backgroundColor: isVerifying ? '#9CA3AF' : '#000000',
-                  color: 'white',
-                  borderRadius: '0.5rem',
-                  fontWeight: '500',
-                  cursor: isVerifying ? 'not-allowed' : 'pointer',
-                  border: 'none',
-                  transition: 'background-color 0.2s',
-                  fontSize: '16px',
-                }}
-                onMouseEnter={(e) => {
-                  if (!isVerifying) {
-                    e.currentTarget.style.backgroundColor = '#1F2937';
-                  }
-                }}
-                onMouseLeave={(e) => {
-                  if (!isVerifying) {
-                    e.currentTarget.style.backgroundColor = '#000000';
-                  }
-                }}
-              >
-                {isVerifying ? (
-                  <div
-                    style={{
-                      display: 'flex',
-                      alignItems: 'center',
-                      justifyContent: 'center',
-                    }}
-                  >
-                    <div
-                      style={{
-                        animation: 'spin 1s linear infinite',
-                        borderRadius: '50%',
-                        height: '1rem',
-                        width: '1rem',
-                        border: '2px solid transparent',
-                        borderBottomColor: 'white',
-                        marginRight: '0.5rem',
-                      }}
-                    ></div>
-                    처리 중...
-                  </div>
-                ) : (
-                  '인증번호 확인'
-                )}
-              </button>
-
-              {/* 재발송 버튼 */}
-              <button
-                type="button"
-                onClick={handleResendCode}
-                disabled={isLoading || countdown > 0}
-                style={{
-                  width: '100%',
-                  height: '40px',
-                  backgroundColor: 'transparent',
-                  color: countdown > 0 ? '#9CA3AF' : '#000000',
-                  borderRadius: '0.5rem',
-                  fontWeight: '500',
-                  cursor: countdown > 0 ? 'not-allowed' : 'pointer',
-                  border: '1px solid #D1D5DB',
-                  transition: 'all 0.2s',
-                  fontSize: '14px',
-                }}
-                onMouseEnter={(e) => {
-                  if (countdown === 0) {
-                    e.currentTarget.style.backgroundColor = '#F9FAFB';
-                  }
-                }}
-                onMouseLeave={(e) => {
-                  if (countdown === 0) {
-                    e.currentTarget.style.backgroundColor = 'transparent';
-                  }
-                }}
-              >
-                {isLoading ? '처리 중...' : '인증번호 재발송'}
-              </button>
-
-              {/* 1/2 단계로 돌아가기 */}
-              <button
-                type="button"
-                onClick={handleBackToStep1}
-                style={{
-                  background: 'none',
-                  border: 'none',
-                  padding: 0,
-                  fontSize: '14px',
-                  color: '#6B7280',
-                  textDecoration: 'underline',
-                  cursor: 'pointer',
-                  fontFamily: 'inherit',
-                }}
-              >
-                다른 이메일로 인증하기
-              </button>
-            </form>
-          )}
-
-          {/* 에러 메시지 */}
-          {error && (
-            <div
-              style={{
-                fontSize: '14px',
-                color: '#EF4444',
-                textAlign: 'center',
-                marginTop: '1rem',
-              }}
-            >
-              {error}
-            </div>
-          )}
-
-          {/* 인증 에러 메시지 */}
-          {verificationError && (
-            <div
-              style={{
-                fontSize: '14px',
-                color: '#EF4444',
-                textAlign: 'center',
-                marginTop: '1rem',
-              }}
-            >
-              {verificationError}
-            </div>
-          )}
-        </div>
-      </div>
-    </div>
+    <EmailVerificationSection
+      email={email}
+      setEmail={setEmail}
+      isLoading={isLoading}
+      onSubmit={handleEmailSubmit}
+      error={error}
+    />
   );
 }
